third_party/blink/web_tests/external/wpt/focus/focus-restoration-in-different-site-iframes-window-expected.txt

This is a testharness.js-based test.
[FAIL] Check result
  assert_equals: Check log expected "outerlog:log:willfocuswindow,windowfocus,didfocuswindow,windowblur,windowfocus," but got "outerlog:log:willfocuswindow,didfocuswindow,"
Harness: the test ran to completion.
